+++ connect/api/v2/organizations/views.py
@@ -13,6 +13,7 @@ from connect.api.v1.organization.permissions import (
     Has2FA,
     OrganizationHasPermission,
 )
+
 from connect.api.v2.organizations.serializers import (
     OrganizationSeralizer,
     NestedAuthorizationOrganizationSerializer
@@ -21,6 +22,9 @@ from connect.api.v2.projects.serializers import ProjectSerializer
 from connect.api.v2.organizations.api_schemas import (
     create_organization_schema,
 )
+
+from connect.api.v2.paginations import CustomCursorPagination
+
 
 
 class OrganizationViewSet(
@@ -35,6 +39,7 @@ class OrganizationViewSet(
     serializer_class = OrganizationSeralizer
     lookup_field = "uuid"
     permission_classes = [IsAuthenticated, OrganizationHasPermission, Has2FA]
+    pagination_class = CustomCursorPagination
 
     def get_queryset(self, *args, **kwargs):
         if getattr(self, "swagger_fake_view", False):